etl/requirements.txt
backoff==2.2.1
clickhouse-driver==0.2.9
kafka-python==2.0.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
requests==2.32.3
